Modularizing matUtils.
